requirements.txt
streamlit
numpy==1.26.4
opencv-python
scikit-image
Pillow